singbox_manager/requirements.txt
PyYAML>=6.0.0
paramiko>=3.0.0
scp>=0.14.0
